Input lowercase without spaces ciphertext to get the plaintext!
The code written in C++ takes in input lowercase ciphertext without any
spaces. It then guesses the key length and if guessed incorrectly allows
the user to input the correct key length. It then guesses the key and
outputs the plaintext (lowercase and without spaces).
This is my solution to the assignment given to me in Applied
Cryptography Course (Monsoon'15, IIIT-Delhi)